# backend/openapi_milestone4.yaml
openapi: 3.0.3
info:
  title: PulseTrack HRMS API - Milestone 4
  version: 1.0.0
  description: |
    # PulseTrack HRMS - GenAI-Powered Human Resource Management System
    
    ## Project Information
    - **Team**: Team 11, IIT Madras
    - **Course**: Software Engineering (SEP 2025)
    - **Milestone**: 4 - API Endpoints Documentation
    
    ## Overview
    PulseTrack HRMS is a comprehensive, AI-powered Human Resource Management System designed to revolutionize 
    workforce management by making HR processes smarter, faster, and more human-centric.
    
    ## GenAI Integrations
    This system integrates multiple GenAI services to automate and enhance HR processes:
    
    ### 1. Google Gemini API
    - **Purpose**: Natural language processing, content generation, and intelligent analysis
    - **Used In**: Policy RAG, Resume Screening, Job Description Generation
    - **Model**: gemini-1.5-flash
    - **Features**: Text generation, embedding, multi-turn conversations
    
    ### 2. ChromaDB
    - **Purpose**: Vector database for semantic search
    - **Used In**: Policy RAG for document similarity search
    - **Features**: Persistent storage, similarity search, metadata filtering
    
    ### 3. Langchain
    - **Purpose**: RAG pipeline orchestration
    - **Used In**: Policy RAG question-answering system
    - **Features**: Document loading, text splitting, retrieval chains
    
    ### 4. PyPDF2
    - **Purpose**: PDF text extraction
...
    
...
    
...
    
...
    
...
    
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from /api/v1/auth/login endpoint"
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/employees:
    get:
      tags:
        - Employee Management
      summary: Get All Employees (Paginated)
      description: |
        **User Story #4: Employee Database**
        Maintain and update employee personal, job, and contact details in one central place.
        
        **Features:**
        - Pagination support
        - Search by name, email, or employee ID
        - Filter by department, team, role, active status
        - Returns summary view of employees
        
        **Error Handling:**
        - 403: HR access required
        - 422: Invalid query parameters
        
        **Access:** HR only
      operationId: getEmployees
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, email, or employee ID
        - name: department_id
          in: query
          schema:
            type: integer
        - name: team_id
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
            enum: [employee, manager, hr]
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  employees:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_pages:
                    type: integer
        '403':
          description: Forbidden - HR access required
    
    post:
      tags:
        - Employee Management
      summary: Create New Employee
      description: |
        **User Story #4: Employee Database**
        Create a new employee account with all required information.
        
        **Features:**
        - Auto-generate employee ID if not provided
        - Hash password securely
        - Set initial leave balances
        - Validate department, team, manager
        - Email uniqueness validation
        
        **Access:** HR only
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                position:
                  type: string
                department_id:
                  type: integer
                team_id:
                  type: integer
                manager_id:
                  type: integer
                role:
                  type: string
                  enum: [employee, manager, hr]
              required:
                - name
                - email
                - password
      responses:
        '201':
          description: Employee created successfully
        '400':
          description: Bad Request - Validation failed
        '403':
          description: Forbidden - HR access required
...
  /api/v1/jobs:
    get:
      tags:
        - Jobs
      summary: Get All Job Listings
      description: |
        **User Story #3: Job Posting Management**
        Manage all job postings in centralized portal to view, edit, or close listings.
        
        **Features:**
        - View all job listings
        - Filter by status, department, location
        - Search by title or description
        
        **Access:** HR (all jobs), Others (active jobs only)
      operationId: getJobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, closed]
        - name: department
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Job listings retrieved successfully
    
    post:
      tags:
        - Jobs
      summary: Create Job Listing
      description: |
        **User Stories #2 & #3: Job Description Management, Job Posting Management**
        Create a new job listing (can be from AI-generated JD or manual).
        
        **Access:** HR only
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                department:
                  type: string
                location:
                  type: string
                job_type:
                  type: string
                status:
                  type: string
                  enum: [draft, active, closed]
              required:
                - title
                - description
      responses:
        '201':
          description: Job listing created successfully
        '403':
          description: Forbidden - HR access required
  /api/v1/feedback:
    post:
      tags:
        - Feedback
      summary: Create Feedback
      description: |
        **User Story #14: Employee Reviews**
        Review team employees to provide continuous feedback and help them improve.
        
        **Features:**
        - Create feedback for employees
        - Support for different feedback types (positive, constructive, goal-related)
        - Optional 1-5 rating
        - Feedback visible to employee and HR
        
        **Access:** Manager or HR only
      operationId: createFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employee_id:
                  type: integer
                feedback_type:
                  type: string
                  enum: [positive, constructive, goal-related, performance, general]
                content:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
              required:
                - employee_id
                - content
      responses:
        '201':
          description: Feedback created successfully
        '403':
          description: Forbidden - Manager/HR access required
...
  /api/v1/leaves:
    post:
      tags:
        - Leave Management
      summary: Create Leave Request
      description: |
        **User Story #9: Leave Notifications**
        Submit a leave request. Notifications will be sent upon approval/rejection.
        
        **Access:** All authenticated users
      operationId: createLeaveRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                leave_type:
                  type: string
                  enum: [casual, sick, annual, maternity, paternity]
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                reason:
                  type: string
              required:
                - leave_type
                - start_date
                - end_date
      responses:
        '201':
          description: Leave request created successfully
        '400':
          description: Bad Request - Insufficient leave balance
  /api/v1/goals:
    post:
      tags:
        - Goals
      summary: Create Goal
      description: |
        **User Story #12: Goal Setting**
        Set clear and effective goals for employees to improve their performance.
        
        **Features:**
        - Create goals for team members
        - Set target dates and success criteria
        - Track progress
        
        **Access:** Manager or HR only
      operationId: createGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employee_id:
                  type: integer
                title:
                  type: string
                description:
                  type: string
                target_date:
                  type: string
                  format: date
                success_criteria:
                  type: string
              required:
                - employee_id
                - title
      responses:
        '201':
          description: Goal created successfully
        '403':
          description: Forbidden - Manager/HR access required
...